Switch from nopts to minimist for args parsing
It turns out that `yo` CLI uses `minimist` under the hood, even though
`yeoman-generator` itself uses `nopts`.

The notable difference is in how an unknown option with a value is
handled. Consider the following command:

  lb export-api-def -o ./testApp.yaml

nopts treats "-o" as a boolean flag, leaving "./testApp.yaml" as an
unprocessed argument.

minimist recognizes "-o" as a flag with value "./testApp.yaml", which
is exactly what we need here.